+++ src/AgentAcademy.Server/Program.cs
@@ -244,6 +244,8 @@ builder.Services.AddSingleton<ICommandHandler, SetPlanHandler>();
 builder.Services.AddSingleton<ICommandHandler, AddTaskCommentHandler>();
 builder.Services.AddSingleton<ICommandHandler, RecallAgentHandler>();
 builder.Services.AddSingleton<ICommandHandler, CloseRoomHandler>();
+builder.Services.AddSingleton<ICommandHandler, CreateRoomHandler>();
+builder.Services.AddSingleton<ICommandHandler, ReopenRoomHandler>();
 builder.Services.AddSingleton<ICommandHandler, MergeTaskHandler>();
 builder.Services.AddSingleton<ICommandHandler, CancelTaskHandler>();
 builder.Services.AddSingleton<ICommandHandler, ShellCommandHandler>();